move `resolve_files` and `resolve_globs` to `scanner/auto_content`
