+++ crates/oxide/src/scanner/mod.rs
@@ -1,1 +1,2 @@
 pub mod allowed_paths;
+pub mod auto_content;
